+++ src/ecs/weaponcomponent.cpp
@@ -28,7 +28,7 @@ void WeaponComponent::update()
         for(int i=0; i<12; i++)
         {
             auto& e = entity->m_manager.enqueueEntity();
-            double angle = rand() % 30;
+            int angle = rand() % 30;
             angle *= std::pow(-1, i);
             Vector2D randpoint = Math2D::rotate_point(transform->center(), angle, {mousePt.x, mousePt.y});
             e.addComponent<ProjectileComponent>(
